feat: cardList 페이지 hook으로 분리
